feat: organize folders and add route to remove account page
